backend/requirements.txt
fastapi>=0.115.8, <0.120.0
uvicorn>=0.24.0, <0.26.0
supabase>=2.4.0, <3.0.0
asyncpg>=0.29.0, <0.31.0
python-multipart>=0.0.6, <0.1.0
pytesseract>=0.3.10, <0.4.0
pillow>=10.0.0, <12.0.0
opencv-python>=4.8.0, <5.0.0
python-dotenv>=1.0.0, <2.0.0
httpx>=0.26.0, <0.28.0
